Restructuring the Project
