+++ AMS/ViewModels/MainViewModel.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AMS.ViewModels
+{
+    class MainViewModel:ViewModelBase
+    {
+        public ViewModelBase CurrentViewModel { get; }
+
+        public MainViewModel()
+        {
+            CurrentViewModel = new HomeViewModel();
+        }
+    }
+}
