--- AMS/Views/AMSView.xaml
@@ -1,37 +1,0 @@
-﻿<UserControl
-    x:Class="AMS.Views.AMSView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:AMS.Views"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewssubcomponents="clr-namespace:AMS.ViewsSubComponents"
-    d:DesignHeight="450"
-    d:DesignWidth="800"
-    mc:Ignorable="d">
-
-    <Grid x:Name="Child_Grid" ShowGridLines="false">
-
-        <Grid.ColumnDefinitions>
-
-            <ColumnDefinition Width="0.5*" />
-
-            <ColumnDefinition Width="0.5*" />
-
-        </Grid.ColumnDefinitions>
-
-        <viewssubcomponents:DirectoryManager
-            x:Name="DirectoryManager_Component"
-            Grid.Column="0"
-            Margin="0,0,5,0"
-            DataContext="{Binding directoryManager_ViewModel}" />
-
-        <viewssubcomponents:DocxModifier
-            x:Name="DocxModifier_Component"
-            Grid.Column="1"
-            Margin="5,0,0,0"
-            DataContext="{Binding docxModifier_ViewModel}" />
-
-    </Grid>
-
-</UserControl>
